requirements.txt
numpy==1.24.0
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
backports-datetime-fromisoformat==2.0.3
braceexpand==0.1.7
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
cytoolz==1.0.1
datasets==2.14.4
decorator==5.2.1
dill==0.3.7
Distance==0.1.3
distro==1.9.0
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
fiddle==0.3.0
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
g2p-en==2.1.0
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
inflect==7.5.0
intervaltree==3.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jiwer==4.0.0
joblib==1.5.2
kaldi-python-io==1.2.2
kaldiio==2.18.1
kiwisolver==1.4.9
lazy_loader==0.4
Levenshtein==0.27.1
lhotse==1.30.3
libcst==1.8.2
librosa==0.11.0
lightning==2.4.0
lightning-utilities==0.15.2
lilcom==1.8.1
llvmlite==0.44.0
loguru==0.7.3
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.1.6
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.15
nemo-toolkit==2.2.1
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
onnx==1.17.0
openai==1.105.0
openai-whisper
optuna==4.5.0
packaging==24.2
pandas==2.3.2
parso==0.8.5
peft==0.17.1
pexpect==4.9.0
pillow==11.3.0
plac==1.4.5
platformdirs==4.4.0
pooch==1.8.2
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyarrow==18.0.0
pybind11==3.0.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyloudnorm==0.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.4
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.0
regex==2025.9.1
requests==2.32.5
resampy==0.4.3
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
sacremoses==0.1.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.35.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==0.5.0.post1
SQLAlchemy==2.0.43
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
text-unidecode==1.3
texterrors==0.5.1
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
tomli==2.2.1
toolz==1.0.0
torch==2.8.0
torchaudio==2.8.0
torchmetrics==1.8.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
triton==3.4.0
typeguard==4.4.4
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.3
wcwidth==0.2.13
webdataset==1.0.2
Werkzeug==3.1.3
wget==3.2
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
